# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CommentCreate": {
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "mentions": {
            "items": {
              "type": "integer"
            },
            "maxItems": 50,
            "title": "Mentions",
            "type": "array"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "segment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Id"
          }
        },
        "required": [
          "body"
        ],
        "title": "CommentCreate",
        "type": "object"
      },
      "CommentOut": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/UserRef"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_deleted": {
            "title": "Is Deleted",
            "type": "boolean"
          },
          "is_edited": {
            "title": "Is Edited",
            "type": "boolean"
          },
          "is_resolved": {
            "title": "Is Resolved",
            "type": "boolean"
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/MentionRef"
            },
            "title": "Mentions",
            "type": "array"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "replies": {
            "items": {
              "$ref": "#/components/schemas/CommentOut"
            },
            "title": "Replies",
            "type": "array"
          },
          "segment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Id"
          },
          "start_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ms"
          }
        },
        "required": [
          "id",
          "segment_id",
          "parent_id",
          "start_ms",
          "author",
          "body",
          "mentions",
          "is_resolved",
          "is_deleted",
          "is_edited",
          "created_at",
          "replies"
        ],
        "title": "CommentOut",
        "type": "object"
      },
      "CommentUpdate": {
        "description": "Partial update. Every field optional; unset fields are left alone.",
        "properties": {
          "body": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "is_resolved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Resolved"
          }
        },
        "title": "CommentUpdate",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable identifier; safe to branch on.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "Field-level errors, keyed by dotted path, when applicable.",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human-readable explanation. Wording may change.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Every non-2xx response body in the API has this shape.\n\nNested under a single `error` key rather than flattened, so a client can\ntell an error from a successful payload without inspecting the status code \u2014\nwhich matters for the fetch wrapper in T-06.4.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MentionRef": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "participant_id": {
            "title": "Participant Id",
            "type": "integer"
          }
        },
        "required": [
          "participant_id",
          "display_name"
        ],
        "title": "MentionRef",
        "type": "object"
      },
...
      "Page_CommentOut_": {
        "properties": {
          "has_next": {
            "readOnly": true,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "1-indexed page number.",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total matching rows, ignoring pagination.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "readOnly": true,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages",
          "has_next"
        ],
        "title": "Page[CommentOut]",
        "type": "object"
      },
...
      "UserRef": {
        "description": "The minimum needed to render an avatar and a name.\n\nDeliberately not the full user record: this is embedded in every meeting\nrow, so anything added here is paid for 20 times per page.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UserRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nMeetings, transcripts, AI summaries and action items.\n\n**Conventions**\n\n* Every list endpoint returns the same envelope:\n  `{items, page, page_size, total, total_pages, has_next}`.\n* Every error returns `{error: {code, message, details}}`. Branch on `code`,\n  which is stable; `message` is for humans and may be reworded.\n* Times are UTC ISO-8601. Positions within a recording are integer\n  **milliseconds** (`start_ms`), never formatted strings.\n* A deleted meeting returns **410**, not 404 \u2014 it is restorable, and the client\n  can offer that.\n",
    "title": "Fireflies API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/comments/{comment_id}": {
      "delete": {
        "description": "Author-only, soft. A deleted parent with replies survives as a 'Comment deleted' tombstone so the thread doesn't collapse.",
        "operationId": "delete_comment_api_v1_comments__comment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error. `details.request_id` correlates to logs."
          }
        },
        "summary": "Delete a comment",
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "description": "Author-only. Body edits set the `edited` marker; resolve applies to threads.",
        "operationId": "update_comment_api_v1_comments__comment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error. `details.request_id` correlates to logs."
          }
        },
        "summary": "Edit or resolve a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
    "/api/v1/meetings/{meeting_id}/comments": {
      "get": {
        "description": "Top-level threads in timeline order; replies nested one level.",
        "operationId": "list_comments_api_v1_meetings__meeting_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          },
          {
            "description": "1-indexed page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-indexed page number.",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page. Clamped to 100.",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page. Clamped to 100.",
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CommentOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No meeting with this id."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Deleted, but restorable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error. `details.request_id` correlates to logs."
          }
        },
        "summary": "List comment threads",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "description": "Optionally anchored to a segment (`segment_id`) or replying to a top-level comment (`parent_id`). Mentions are participant ids.",
        "operationId": "create_comment_api_v1_meetings__meeting_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No meeting with this id."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Deleted, but restorable."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error. `details.request_id` correlates to logs."
          }
        },
        "summary": "Add a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
  }
}